Rails 6 compatibility.
First step is to remove dependency on arel gem as it is part of
active_record now
