--- lib/postgres_ext/active_record/5.0/relation.rb
@@ -1,9 +1,0 @@
-module ActiveRecord
-  class Relation
-    class WhereClause
-      def modified_predicates(&block)
-        WhereClause.new(predicates.map(&block), binds)
-      end
-    end
-  end
-end
