+++ lib/postgres_ext/active_record/query_methods.rb
