+++ lib/postgres_ext/active_record/querying.rb
