--- lib/postgres_ext/arel/4.1/visitors.rb
@@ -1,2 +1,0 @@
-require 'postgres_ext/arel/4.1/visitors/depth_first'
-require 'postgres_ext/arel/4.1/visitors/postgresql'
