--- lib/postgres_ext/arel/4.1/visitors/depth_first.rb
@@ -1,9 +1,0 @@
-require 'arel/visitors/depth_first'
-
-module Arel
-  module Visitors
-    class DepthFirst
-      alias :visit_IPAddr :terminal
-    end
-  end
-end
