--- lib/postgres_ext/arel/4.2/visitors.rb
@@ -1,1 +1,0 @@
-require 'postgres_ext/arel/4.2/visitors/postgresql'
