--- lib/postgres_ext/arel/5.0/visitors.rb
@@ -1,1 +1,0 @@
-require 'postgres_ext/arel/5.0/visitors/postgresql'
