+++ lib/postgres_ext/arel/nodes/array_nodes.rb
@@ -1,5 +1,3 @@
-require 'arel/nodes/binary'
-
 module Arel
   module Nodes
     class Overlap < Arel::Nodes::Binary
