+++ lib/postgres_ext/arel/nodes/contained_within.rb
@@ -1,4 +1,3 @@
-require 'arel/nodes/binary'
 module Arel
   module Nodes
     class ContainedWithin < Arel::Nodes::Binary
