+++ lib/postgres_ext/arel/visitors.rb
@@ -1,0 +1,1 @@
+require 'postgres_ext/arel/visitors/postgresql'
