+++ lib/postgres_ext/arel/visitors/postgresql.rb
@@ -1,5 +1,3 @@
-require 'arel/visitors/postgresql'
-
 module Arel
   module Visitors
     class PostgreSQL
@@ -14,7 +12,7 @@ module Arel
       end
 
       def visit_Arel_Nodes_Contains o, collector
-        left_column = o.left.relation.engine.columns.find do |col|
+        left_column = o.left.relation.name.classify.constantize.columns.find do |col|
           col.name == o.left.name.to_s || col.name == o.left.relation.name.to_s
         end
