+++ postgres_ext.gemspec
@@ -16,8 +16,7 @@ Gem::Specification.new do |gem|
   gem.require_paths = ["lib"]
   gem.version       = PostgresExt::VERSION
 
-  gem.add_dependency 'activerecord', '>= 4.0.0'
-  gem.add_dependency 'arel', '>= 4.0.1'
+  gem.add_dependency 'activerecord', '>= 6.0.0'
   gem.add_dependency 'pg_array_parser', '~> 0.0.9'
 
   gem.add_development_dependency 'rake', '~> 10.1.0'
